Added gallery mode
Added gallery mode and bound it to 'g' key. When g is pressed, a small slide show of saved pictures will play on loop until the g key is pressed again to close it and play will resume. Currently it opens in another window. Here is a link to article that explains more about the code used: https://dev.to/siddharth2016/how-to-make-a-slideshow-using-opencv-5ge7